backend/requirements.txt
cloudinary==1.43.0
dotenv==0.9.9
firebase-admin==6.9.0
Flask==3.1.0
flask-cors==6.0.1
GeoAlchemy2==0.17.1
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
oauthlib==3.2.2
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
shapely==2.1.1
SQLAlchemy==2.0.39
urllib3==2.3.0
